fix: fixed react errors/warnings
